# project-notes/api-spec.yaml
openapi: 3.1.0
info:
  title: Project Allocation Management System (PAMS) API
  version: 1.12.0
  description: |
    REST API for the Project Allocation Management System.
    All endpoints require a Keycloak-issued OIDC Bearer token unless stated otherwise.
    Obtain tokens via the Keycloak Authorization Code flow at the configured realm (`/realms/pams`).
    Identity resolution uses the `empCode` JWT claim for Keycloak-to-PAMS identity bridging (sub claim verified against employees table with empCode fallback).
    Dates use ISO 8601 `YYYY-MM-DD` format. Timestamps use UTC (`YYYY-MM-DDTHH:MM:SSZ`).

    **Pagination (NFR-18/19):** All list endpoints support `page` (default 1) and `limit` (default 10, max 100).
    **Single-endpoint strategy (NFR-20):** Each resource has one list endpoint that also supports search and filter via query parameters.
    **Role-based response shaping (NFR-21):** Responses are filtered server-side based on the authenticated user's role.
    **Consolidated REST design:** Deactivation is handled via PUT (state change), allocation stop via PATCH, allocation remove via DELETE (soft-delete).
    **Billable flag & project status:** Projects carry a `billable` boolean (defaults to true for Client accounts) and a lifecycle `status` (Upcoming → Active → Completed). Both HR and PM (scoped to own projects) can modify status.
...
  contact:
    name: PAMS Engineering
  license:
    name: Internal Use Only
servers:
  - url: https://api.pams.internal/api/v1
...
security:
  - BearerAuth: []
...
paths:
...
  /allocations:
    get:
      tags: [Allocations]
      summary: List allocations (paginated, filterable)
      description: |
        Paginated list of allocations with optional filters.
        All authenticated users can call this endpoint; role-based scoping is applied in the handler
        (e.g. Staff see only their own allocations, PM see allocations on their projects).
      operationId: allocations_list
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: empCode
          in: query
          schema:
            type: string
          description: Filter by allocated employee code.
        - name: projectCode
          in: query
          schema:
            type: string
          description: Filter by project code.
        - name: projectManagerEmpCode
          in: query
          schema:
            type: string
          description: Filter allocations on projects managed by this PM (by empCode). Use with current user's empCode to get "my managed project allocations".
        - name: status
          in: query
          schema:
            type: string
            enum: [Active, Ended, Upcoming]
          description: Filter by computed allocation status.
        - name: billable
          in: query
          schema:
            type: boolean
          description: Filter by resource-level billable flag.
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Paginated list of allocations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponse_AllocationDetail"
        "400":
          $ref: "#/components/responses/InvalidSortError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /allocations/export:
    post:
      tags: [Allocations]
      summary: Export allocations as PDF or Excel (HR, PM)
      description: |
        Returns all matching allocations (no pagination) as a downloadable PDF or Excel file.
        Accepts the same filter and sort query parameters as the list endpoint.
        Role-based scoping applies (Staff → own, PM → own projects, HR → all).
        An optional JSON body may specify a column name map to control which columns appear
        in the export and their display labels.
      operationId: allocations_export
      parameters:
        - name: empCode
          in: query
          schema:
            type: string
          description: Filter by allocated employee code.
        - name: projectCode
          in: query
          schema:
            type: string
          description: Filter by project code.
        - name: projectManagerEmpCode
          in: query
          schema:
            type: string
          description: Filter allocations on projects managed by this PM.
        - name: status
          in: query
          schema:
            type: string
            enum: [Active, Ended, Upcoming]
        - name: billable
          in: query
          schema:
            type: boolean
        - $ref: "#/components/parameters/SortParam"
        - $ref: "#/components/parameters/ExportExtParam"
      requestBody:
        required: false
        description: |
          Optional column mapping. Keys are response DTO field names (case-insensitive),
          values are display labels for the export header. Only matched columns are included.
          If omitted or empty, all default columns are exported.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportColumnMap"
            example:
              empCode: "Employee Code"
              empName: "Employee"
              projectCode: "Project"
              percentage: "Allocation %"
              status: "Status"
      responses:
        "200":
          description: File download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
              example: 'attachment; filename="allocations-2026-03-05.pdf"'
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [Allocations]
      summary: Create a new allocation (HR, PM)
      description: |
        HR can allocate to any active project.
        PM can only allocate to projects they manage.
        Allocation % must be ≥ minAllocationPct (default 25) and a multiple of allocationIncrement (default 5).
        Total overlapping allocations for the employee must not exceed 100%.
        **Validation:** `fromDate` must not be a past date (>= today). Returns 400 if violated.
      operationId: allocations_create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAllocationRequest"
            example:
              empCode: EMP042
              projectCode: PROJ-ALPHA
              fromDate: "2026-03-01"
              toDate: "2026-06-30"
              percentage: 50
      responses:
        "201":
          description: Allocation created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationDetailResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Forbidden (caller is PM but does not manage target project)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://pams.internal/errors/ERR_NOT_PROJECT_OWNER"
                title: Forbidden
                status: 403
                detail: You are not the Project Manager of this project.
        "422":
          description: Capacity exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapacityExceededProblem"
              example:
                type: "https://pams.internal/errors/ERR_CAPACITY_EXCEEDED"
                title: Capacity Exceeded
                status: 422
                detail: "Employee EMP042 is already 75% allocated from 2026-03-01. Available: 25%."
                extensions:
                  empCode: EMP042
                  conflictFrom: "2026-03-01"
                  currentTotal: 75
                  requested: 50
                  available: 25
  /allocations/{allocationId}:
    parameters:
      - $ref: "#/components/parameters/AllocationId"

    get:
      tags: [Allocations]
      summary: Get allocation by ID
      operationId: allocations_getById
      responses:
        "200":
          description: Allocation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Allocations]
      summary: Update allocation percentage or date range (HR, PM)
      description: |
        PM can only update allocations on their own projects.
        All capacity rules apply (excluding the current allocation from the check).
        Cannot edit ended allocations.
      operationId: allocations_update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAllocationRequest"
      responses:
        "200":
          description: Allocation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationDetailResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Capacity exceeded or allocation already ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

    patch:
      tags: [Allocations]
      summary: Stop an allocation (HR, PM)
      description: |
        Stops an active or upcoming allocation.
        If allocation has not yet started (fromDate > today): sets toDate = fromDate (satisfies chk_allocation_dates constraint).
        If allocation is active (fromDate ≤ today): sets toDate = tomorrow.
        PM can only stop allocations on their own projects.
        Request body: `{"action": "stop"}`.
      operationId: allocations_stop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [stop]
            example:
              action: stop
      responses:
        "200":
          description: Allocation stopped. Returns updated allocation with new toDate.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Allocation already ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://pams.internal/errors/ERR_ALLOCATION_ALREADY_ENDED"
                title: Unprocessable Entity
                status: 422
                detail: This allocation has already ended.

    delete:
      tags: [Allocations]
      summary: Soft-delete a past allocation (HR, PM)
      description: |
        Only ended allocations (toDate < today) can be removed.
        Sets deletedAt timestamp. Record is hidden from default views.
        HR can see removed allocations via includeRemoved=true on list endpoints.
      operationId: allocations_remove
      responses:
        "200":
          description: Allocation removed (soft-deleted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Cannot remove active or future allocation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://pams.internal/errors/ERR_ALLOCATION_NOT_ENDED"
                title: Unprocessable Entity
                status: 422
                detail: Only ended allocations can be removed.
  /allocations/capacity-check:
    get:
      tags: [Allocations]
      summary: Check allocation capacity for an employee in a date window
      description: Returns the current total allocated % and available % for the employee in the specified window.
      operationId: allocations_capacityCheck
      parameters:
        - name: empCode
          in: query
          required: true
          schema:
            type: string
          description: Employee code to check capacity for.
        - name: fromDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
          description: Null for open-ended check.
        - name: excludeAllocationId
          in: query
          schema:
            type: string
            format: uuid
          description: Exclude this allocation ID when checking (used during edit).
      responses:
        "200":
          description: Capacity information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapacityCheckResponse"
              example:
                empCode: EMP042
                windowFrom: "2026-03-01"
                windowTo: "2026-06-30"
                currentTotalPercentage: 50
                availablePercentage: 50
                allocationStatus: Partial
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── SYSTEM CONFIG ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC Bearer token issued by Keycloak (RS256 signed).
        Obtain via Keycloak Authorization Code flow: `{Authority}/protocol/openid-connect/auth`
        or Client Credentials flow for machine-to-machine.
        Default realm: `https://keycloak.internal/realms/pams`.

  # ── PARAMETERS ─────────────────────────────────────────────────────────────
  parameters:
...
    AllocationId:
      name: allocationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: "Number of items per page (default: 10, max: 100). Per NFR-18."
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: |
        Sort field. Use `fieldName` for ascending or `-fieldName` for descending.
        Each endpoint has a default sort if omitted. Only one sort field per request.
        Invalid field names return 400 with a ProblemDetails response containing
        `field` (the invalid field) and `allowedFields` (list of valid field names).
      example: "-fromDate"
    ExportExtParam:
      name: ext
      in: query
      required: true
      schema:
        type: string
        enum: [pdf, xls]
      description: "Export format: `pdf` for PDF or `xls` for Excel (.xlsx)."

  # ── RESPONSES ──────────────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://pams.internal/errors/ERR_UNAUTHORIZED"
            title: Unauthorized
            status: 401
            detail: Authentication token is missing or invalid.
...
    InvalidSortError:
      description: "Invalid sort field. Returns the field name and list of allowed sort fields."
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/InvalidSortProblem"
          example:
            type: "https://pams.internal/errors/ERR_INVALID_SORT"
            title: "Invalid Sort Parameter"
            status: 400
            detail: "Invalid sort field 'xyz'. Allowed fields: accountCode, accountName, accountType, isActive"
            instance: "/api/v1/accounts"
            field: "xyz"
            allowedFields: ["accountCode", "accountName", "accountType", "isActive"]

  # ── SCHEMAS ────────────────────────────────────────────────────────────────
  schemas:
...
    PagedResponse_AllocationDetail:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AllocationDetailResponse"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"

    # ── System Config ──────────────────────────────────────────────────────
...